Откорректировал некоторые моменты, которые мне показались наиболее важными
