--- src/main/java/CreatorListObject/PoloniexDao.java
@@ -1,12 +1,0 @@
-package CreatorListObject;
-
-import java.io.IOException;
-import java.util.List;
-
-public interface PoloniexDao {
-    void load() throws IOException;
-
-    List<JsonPoloniexObject> getPoloniexObjectList();
-
-    JsonPoloniexObject getTokenName(String namePair);
-}
