--- src/main/java/ObjectDB/ConnectionHolder.java
@@ -1,32 +1,0 @@
-package ObjectDB;
-
-import ObjectDB.DbSystemException.DbException;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-
-public class ConnectionHolder {
-    private static ConnectionHolder connectionHolder;
-
-    public static ConnectionHolder getInstance() {
-        if (connectionHolder == null) {
-            connectionHolder = new ConnectionHolder();
-        }
-        return connectionHolder;
-    }
-
-    public Connection getConnection() throws Exception {
-        try {
-            LoaderProperties loadProperties = LoaderProperties.getInstance();
-            return DriverManager.getConnection(
-                    loadProperties.getJdbcUrl(),
-                    loadProperties.getLogin(),
-                    loadProperties.getPassword()
-            );
-
-
-        } catch (Exception e) {
-            throw new DbException("Can't create connection", e);
-        }
-    }
-}
