--- src/main/java/ObjectDB/PoloniexDoaJdbc.java
@@ -1,11 +1,0 @@
-package ObjectDB;
-
-import CreatorListObject.JsonPoloniexObject;
-import ObjectDB.DbSystemException.DbException;
-
-import java.sql.SQLException;
-
-public interface PoloniexDoaJdbc {
-    void insertObject(JsonPoloniexObject jsonPoloniexObject) throws DbException, SQLException;
-    void someSelect() throws DbException, SQLException;
-}
