+++ src/main/java/ru/solovev/fin/statistic/persistence/LoaderProperties.java
@@ -1,0 +1,53 @@
+package ru.solovev.fin.statistic.persistence;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.util.Properties;
+
+public class LoaderProperties {
+
+    private static LoaderProperties loaderProperties;
+
+    //TODO: убери завязку на твой путь, сделай лучше
+    private final String pathToJdbcProperties = "./jdbc.properties";
+    private final String driverClassName;
+    private final String login;
+    private final String password;
+    private final String jdbcUrl;
+
+    public static LoaderProperties getInstance() throws Exception {
+        if (loaderProperties == null) {
+            loaderProperties = new LoaderProperties();
+        }
+        return loaderProperties;
+    }
+
+    private LoaderProperties() throws IOException {
+        Properties properties = new Properties();
+        FileInputStream inputStream = new FileInputStream(pathToJdbcProperties);
+        properties.load(inputStream);
+        inputStream.close();
+
+        this.driverClassName = properties.getProperty("jdbc.driver");
+        this.login = properties.getProperty("jdbc.username");
+        this.password = properties.getProperty("jdbc.password");
+        this.jdbcUrl = properties.getProperty("jdbc.url");
+    }
+
+    public String getDriver() {
+        return driverClassName;
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public String getJdbcUrl() {
+        return jdbcUrl;
+    }
+}
+
